requirements.txt
# Core dependencies
streamlit>=1.28.0
chromadb>=0.4.15
openai>=1.3.0

# Free local embeddings (alternative to OpenAI)
sentence-transformers>=2.2.0
torch>=2.0.0

# Document processing
pdfplumber>=0.10.0
python-docx>=1.1.0
pytesseract>=0.3.10
Pillow>=10.0.0

# HTTP client for Ollama
requests>=2.31.0

# Rate limiting
ratelimit>=2.2.1

# Environment variables
python-dotenv>=1.0.0

# Utilities
tiktoken>=0.5.1
